+++ lib/philiprehberger/cache_kit.rb
@@ -2,6 +2,11 @@
 
 require_relative "cache_kit/version"
 require_relative "cache_kit/entry"
+require_relative "cache_kit/callbacks"
+require_relative "cache_kit/tag_stats"
+require_relative "cache_kit/batch"
+require_relative "cache_kit/serializable"
+require_relative "cache_kit/eviction"
 require_relative "cache_kit/store"
 
 module Philiprehberger